+++ Filters.lua
@@ -2,14 +2,16 @@ local _, addon = ...;
 addon.Filters = {};
 local filters = addon.Filters;
 
-LE_LOOT_FILTER_NEW_RANGE = 100
-LE_LOOT_FILTER_PETS = 101;
-LE_LOOT_FILTER_MOUNTS = 102;
+_G[addon.Metadata.Prefix .. "_LE_LOOT_FILTER_NEW_RANGE"] = 100;
+_G[addon.Metadata.Prefix .. "_LE_LOOT_FILTER_PETS"] = 101;
+_G[addon.Metadata.Prefix .. "_LE_LOOT_FILTER_MOUNTS"] = 102;
+_G[addon.Metadata.Prefix .. "_LE_LOOT_FILTER_TOYS"] = 103;
 
 local defaults = {
 	profile = {
-		HideOwnedPets = false,
-		HideCollectedMounts = false
+		HideCollectedPets = false,
+		HideCollectedMounts = false,
+		HideCollectedToys = false
 	}
 };
 
@@ -27,28 +29,88 @@ function filters:Load()
 end
 
 function filters:Validate(lootFilter, itemId)
-	if self.Pets(itemId) and addon.Filters.db.profile.HideOwnedPets then
-		return (C_PetJournal.GetNumCollectedInfo((select(13, C_PetJournal.GetPetInfoByItemID(itemId))))) == 0;
-	end
+	if lootFilter == _G[addon.Metadata.Prefix .. "_LE_LOOT_FILTER_PETS"] then
+		return self:ValidatePetsOnly(itemId);
+    elseif lootFilter == _G[addon.Metadata.Prefix .. "_LE_LOOT_FILTER_MOUNTS"] then
+		return self:ValidateMountsOnly(itemId);
+    elseif lootFilter == _G[addon.Metadata.Prefix .. "_LE_LOOT_FILTER_TOYS"] then
+		return self:ValidateToysOnly(itemId);
+	else
+		if self.IsPet(itemId) and addon.Filters.db.profile.HideCollectedPets then
+			return not self.IsPetCollected(itemId);
+		end
 
-	if self.Mounts(itemId) and addon.Filters.db.profile.HideCollectedMounts then
-		return not (select(11, C_MountJournal.GetMountInfoByID(C_MountJournal.GetMountFromItem(itemId))));
+		if self.IsMount(itemId) and addon.Filters.db.profile.HideCollectedMounts then
+			return not self.IsMountCollected(itemId);
+		end
+
+		if self.IsToy(itemId) and addon.Filters.db.profile.HideCollectedToys then
+			return not self.IsToyCollected(itemId);
+		end
 	end
 
 	return true;
-	-- if lootFilter == LE_LOOT_FILTER_PETS then
-	-- 	return self.Pets(itemId);
-    -- elseif lootFilter == LE_LOOT_FILTER_MOUNTS then
-    --     return self.Mounts(itemId);
-	-- end
 end
 
-function filters.Pets(itemId)
-	local classId, subclassId = select(12, GetItemInfo(itemId));
-	return classId == Enum.ItemClass.Miscellaneous and subclassId == Enum.ItemMiscellaneousSubclass.CompanionPet;
+do -- Pets
+	function filters:ValidatePetsOnly(itemId)
+		if not self.IsPet(itemId) then
+			return false;
+		end
+		if addon.Filters.db.profile.HideCollectedPets then
+			return not self.IsPetCollected(itemId);
+		end
+		return true;
+	end
+
+	function filters.IsPet(itemId)
+		local classId, subclassId = select(12, GetItemInfo(itemId));
+		return classId == Enum.ItemClass.Miscellaneous and subclassId == Enum.ItemMiscellaneousSubclass.CompanionPet;
+	end
+
+	function filters.IsPetCollected(itemId)
+		return (C_PetJournal.GetNumCollectedInfo((select(13, C_PetJournal.GetPetInfoByItemID(itemId))))) ~= 0;
+	end
 end
 
-function filters.Mounts(itemId)
-	local classId, subclassId = select(12, GetItemInfo(itemId));
-	return classId == Enum.ItemClass.Miscellaneous and subclassId == Enum.ItemMiscellaneousSubclass.Mount;
+do -- Mounts
+	function filters:ValidateMountsOnly(itemId)
+		if not self.IsMount(itemId) then
+			return false;
+		end
+		if addon.Filters.db.profile.HideCollectedMounts then
+			return not self.IsMountCollected(itemId);
+		end
+		return true;
+	end
+
+	function filters.IsMount(itemId)
+		local classId, subclassId = select(12, GetItemInfo(itemId));
+		return classId == Enum.ItemClass.Miscellaneous and subclassId == Enum.ItemMiscellaneousSubclass.Mount;
+	end
+
+	function filters.IsMountCollected(itemId)
+		return (select(11, C_MountJournal.GetMountInfoByID(C_MountJournal.GetMountFromItem(itemId))));
+	end
+end
+
+do -- Toys
+	function filters:ValidateToysOnly(itemId)
+		if not self.IsToy(itemId) then
+			return false;
+		end
+		if addon.Filters.db.profile.HideCollectedToys then
+			return not self.IsToyCollected(itemId);
+		end
+		return true;
+	end
+
+	function filters.IsToy(itemId)
+		itemId = C_ToyBox.GetToyInfo(itemId);
+		return itemId ~= nil;
+	end
+
+	function filters.IsToyCollected(itemId)
+		return PlayerHasToy(itemId);
+	end
 end